Bump typescript version to 3.7-beta
Since we're in the middle of migrating JS to TS, we need to be able to
have both the --declaration and --allowJs flags enabled on the
compiler. Support for both of these options being enabled at the same
time was only introduced in 3.7-beta, hence the upgrade. Otherwise,
we can not use this package from our other TS packages since there
would be no declaration files for the typescript compiler to import