Simplify process of calculating SAID in helpers
In addition to making the method less hacky by removing the need to
instantiate a fake Coordinator interface, it now generates the SAID
based upon the parameter of "getID" of the Coordinator contract,
removing the need to change the method should the "getID" parameter
ever change.

The only thing here that we lose is that we no longer use the
Coordinator interface for it's ABI, inducing a loss of generality for
generating a service agreement id.